==> 01_gathering_data.ipynb <==
# Gathering data

1. 한국은행 금융통화위원회 의사록
2. '금리'를 포함한 연합인포맥스, 연합뉴스, 이데일리
3. 네이버 금융 각 증권사의 채권분석리포트 
4. 일별 콜금리
5. 한국은행 기준금리

# -*- coding: utf-8 -*-

import pandas as pd

from bs4 import BeautifulSoup as bs
from urllib.request import Request
from urllib.request import urlopen
from urllib.request import urlretrieve

import re
import datetime

## 1. 한국은행 금융통화위원회 의사록

2005.01.01 ~ 2020.20.21 의 약 15년치 자료를 모두 수집하기 위해 노력했다. 

.pdf 파일과 .hwp 파일 모두를 제공하는 지금과 달리 2005년 6월 이전엔 한국은행에서 .hwp 파일로만 의사록을 제공하고 사이트의 다운로드 링크 구조도 달랐다는 것을 알 수 있는데, 이를 고려하여 다운로드 받을 수 있도록 코딩하였다. 

bok_mpb_url = "https://www.bok.or.kr/portal/bbs/B0000245/list.do?menuNo=200761"

def get_date_and_title(page_bs):
    result = []
    title_bs_list = page_bs.find_all('span', {'class':'titlesub'})
    
    for item in title_bs_list:
        item = item.text
        date_regex = re.compile(r'\d{4}\.\d+\.\d+')
        date = date_regex.search(item).group()
        date = datetime.datetime.strptime(date, "%Y.%m.%d").date()

        title_regex = re.compile(r'.*\)\(')
        title = title_regex.search(item).group()[:-1]
        
        date_and_title = {}
        date_and_title['date'] = date # 의사록은 하루에 2개 이상 올라오지 않는다고 가정. 
        date_and_title['title'] = title
        
        result.append(date_and_title)
        
    return result

def find_download_link_ancestor_from_elem(elem):
    file_regex = re.compile(r'fileDown')
    if elem.find('a', {'href':file_regex}):
        return elem
    else:
        return find_download_link_ancestor_from_elem(elem.parent)

def download_mpb_minute(title_bs, title, date_string):
    download_link_ancestor = find_download_link_ancestor_from_elem(title_bs)
    
    file_regex = re.compile(r'fileDown')
    aTag_list = download_link_ancestor.find_all(attrs={'href': file_regex})
    href_list = [x.attrs['href'] for x in aTag_list]
    
    file_type_regex = re.compile(r'hwp|pdf')
    file_type_list = download_link_ancestor.find_all(attrs={'title': file_type_regex})
    file_type_list = [file_type_regex.search(x.text).group() for x in file_type_list]
    
    href2file_type = {}
    if file_type_list == []:
        for href in href_list:
            href2file_type[href] = 'hwp'
    else:
        for idx, href in enumerate(href_list):
            href2file_type[href] = file_type_list[idx]
    
    for href in href_list:
        download_link = 'https://www.bok.or.kr' + href
        urlretrieve(download_link, './data_files/BOK_minutes/' + href2file_type[href] + '/' + date_string + '_-_' + title + '.' + href2file_type[href])

def get_mpb_minutes_files(page):
    print("Downloading MPB documents on page {}...".format(page))
    
    page_url = bok_mpb_url + "&pageIndex=" + str(page)
    page_req = urlopen(page_url)
    page_bs = bs(page_req, 'html.parser')

    title_bs_list = page_bs.find_all('span', {'class':'titlesub'}) ##
    date_and_titles = get_date_and_title(page_bs)
    
#     last_date_string = str(date_and_titles[-1]['date'])
    for idx, dict_elem in enumerate(date_and_titles):
        date = dict_elem['date']
        date_string = str(date)
        
        title = dict_elem['title']
        
        download_mpb_minute(title_bs_list[idx], title, date_string)
    return 1

for page in range(1, 32):
    get_mpb_minutes_files(page)

## 2. 네이버금융 각 증권사 채권분석리포트

네이버 금융에 올라온 모든 채권분석 리포트를 가져온다. 하지만 가장 오래된 자료부터 가져와도 최초자료는 2008 ~ 부터 있고 그 이전은 없었다. 

naver_bond_url = "https://finance.naver.com/research/debenture_list.nhn?"

def get_bond_info_and_download(page):
    print("Downloading bond reports on page {}".format(page))
    
    page_url = naver_bond_url + "&page=" + str(page)
    page_req = urlopen(page_url)
    page_bs = bs(page_req, 'html.parser')
    
    reports_bs_list = page_bs.select("div.box_type_m > table.type_1 > tr > td.file")
    reports_bs_list = [x.parent for x in reports_bs_list]
    
    reports_info_list = []
    
    for bs_elem in reports_bs_list:
        td_elements = []
        for child in bs_elem.select('td'):
            td_elements.append(child)
        
        title = td_elements[0].text
        company = td_elements[1].text
        date = datetime.datetime.strptime(td_elements[3].text, "%y.%m.%d").date()
        download_link = td_elements[2].a.attrs['href']
        forbidden_regex = re.compile(r'[\\/\*"\?><\|:]')
        urlretrieve(download_link, './data_files/NAVER_bond_reports/' + str(date) + "_" + company + re.sub(forbidden_regex, "", title) + ".pdf")
        
        info_dic = {}
        info_dic['report_title'] = title
        info_dic['company'] = company
        info_dic['date'] = date
        
        reports_info_list.append(info_dic)
    return reports_info_list

for page in range(1, 122):
    get_bond_info_and_download(page)
