==> 03_preprocessing_bok_call_rates.ipynb <==
# Preprocessing BOK call rates 

한국은행에서 다운받은 call rates 파일을 전처리한다. 
- 빈칸은 ffill
- string object인 date를 datetime 화. 
- 30일 전 콜금리와 비교하여 올랐는지 내렸는지 동일한지 표기 

import pandas as pd
import numpy as np

df = pd.read_csv('./data_files/BOK_4.1.1_market_call_rate_-_daily.csv', encoding='utf-8')
df.head()

df = df.fillna(method='ffill')

df.date = pd.to_datetime(df.date, format='%m/%d/%Y')

df = df.merge(df.assign(date = df.date+pd.Timedelta(days=30)), 
               on='date',
               how='left', suffixes=['', '_30daysago'])

df.dropna(inplace=True)

df['change'] = np.select(
    condlist=[
        (df.call_rate > df. call_rate_30daysago), 
        (df.call_rate == df.call_rate_30daysago),
    ], 
    choicelist=[
        1,
        0,
    ],
    default=-1
)

df

df.to_pickle('./data_files/BOK_call_rates.pkl')
