==> 03_preprocessing_text_data-mp.ipynb <==
# -*- coding: utf-8 -*-
from ekonlpy.sentiment import MPCK

import pandas as pd
import multiprocessing as mp
import pickle

import os, sys
import re

import preprocessing_mp_worker as worker

# 1. BOK minutes text

Jupyter Lab에서의 `multiprocessing` 사용을 위해 main함수인 `preprocessing()`를 제외한 모든 함수는 `preprocessing_mp_worker.py`에 옮겨졌다. 

Code Attribution: 김현진T, Fininsight

### Korean stop words from here: 
https://www.ranks.nl/stopwords/korean

## Test code: 어떤 식으로 파싱되는지 확인. 

source_folder="./data_files/BOK_minutes/pdf2txt/"
output_file="./data_files/BOK_minutes/pkl/"

txt_files = [f for f in os.listdir(source_folder) if os.path.isfile(os.path.join(source_folder, f))]
txt_files = [f for f in txt_files if f[-4:] == '.txt']
txt_files.sort()

len(txt_files)

df = pd.DataFrame(columns=['date', 'minutes'])
df["Economic Situation"] = ""
df["Foreign Currency"] = ""
df["Financial Markets"] = ""
df["Monetary Policy"] = ""
df["Participant Views"] = ""
df["Government View"] = ""

df["Economic Situation count"] = ""
df["Foreign Currency count"] = ""
df["Financial Markets count"] = ""
df["Monetary Policy count"] = ""
df["Participant Views count"] = ""
df["Government View count"] = ""
df['ngrams'] =""

### 한 문서의 ngram의 갯수가 매우 많다. 이를 모두 각각 하나의 row로 만들면 df의 길이가 매우 커진다. 

## 1) 0번째 문서의 모든 ngram을 ,(comma)로 separate한 경우 (원본코드의 방법)

%%time

for txt_file in txt_files[:1]:
    with open(source_folder + txt_file, 'r',encoding='utf-8') as f :
        txt = f.read()
        sections, section_texts = worker.preprocess_minutes(txt)

        print(txt_file)

        ngrams = []
        for i, section_text in enumerate(section_texts[1:3]):
            print("doing #{} iteration...".format(i))
            with mp.Pool(processes=6) as p:
                mp_result = p.map(worker.text2ngram, section_text)
            for i in mp_result:
                ngrams += i

    df.loc[len(df)] = [txt_file.split('_-_')[0],
                       txt,
                       '@@@'.join(section_texts[0]),
                       '@@@'.join(section_texts[1]),
                       '@@@'.join(section_texts[2]),
                       '@@@'.join(section_texts[3]),
                       '@@@'.join(section_texts[4]),
                       '@@@'.join(section_texts[5]),
                       len(section_texts[0]),
                       len(section_texts[1]),
                       len(section_texts[2]),
                       len(section_texts[3]),
                       len(section_texts[4]),
                       len(section_texts[5]),
                       ngrams, # list 형태
                      ]

    

### Performance comparison: list vs str

df에 저장할 때 ngram들이 들어있는 list와 이를 , 로 join한 긴 str 둘 중 어떤 것이 더 나은지 확인해보았다. 

- `','.join(ngrams)`로 Concatenated string으로 저장할 시
    -  Wall time: 12s, 
    - 필수 columns memory sum: 96523, 
    - ngrams memory: 43274

- `ngrams`로 list로 저장할 시 
    - Wall time: 11.8s, 
    - 필수 columns memory sum: 75353, 
    - ngrams memory: 22104
                      
승자는 list. 추후 사용성을 생각하더라도 list 구조가 나을 것으로 보인다. 

df.memory_usage(index=True, deep=True)

df[['date', 'minutes', 'Economic Situation', 'Foreign Currency','ngrams']].memory_usage(index=True, deep=True).sum()

Peek inside ngram results... 

- '엔달러/NNG', '공급우위/NNG', '국제투자은행/NNG', 등을 보면 역시 eKoNLPy 답게 경제단어를 KoNLPy처럼 자르지 않고 잘 추출한 것을 볼 수 있다. 
-  '엔달러/NNG;환율/NNG;상승/NNG', '기대/NNG;원화/NNG;절상/NNG', '통화/NNG;절상/NNG', '전망/NNG;상향/NNG', 등을 보면 5-gram을 했기 때문에 token 묶음이 나오기도 하는 것을 볼 수 있다. ; 로 연결되어있다. 

df.ngrams

## 2) 0번째 문서의 모든 ngram을 각각 row로 만들어 df에 담았을 경우

%%time

for txt_file in txt_files[:1]:
    with open(source_folder + txt_file, 'r',encoding='utf-8') as f :
        txt = f.read()
        sections, section_texts = worker.preprocess_minutes(txt)

        print(txt_file)

        ngrams = []
        for i, section_text in enumerate(section_texts[1:3]):
            print("doing #{} iteration...".format(i))
            with mp.Pool(processes=6) as p:
                mp_result = p.map(worker.text2ngram, section_text)
                for ngram_list in mp_result:
                    for item in ngram_list:
                        df.loc[len(df)] = [txt_file.split('_-_')[0],
                                   txt,
                                   '@@@'.join(section_texts[0]),
                                   '@@@'.join(section_texts[1]),
                                   '@@@'.join(section_texts[2]),
                                   '@@@'.join(section_texts[3]),
                                   '@@@'.join(section_texts[4]),
                                   '@@@'.join(section_texts[5]),
                                   len(section_texts[0]),
                                   len(section_texts[1]),
                                   len(section_texts[2]),
                                   len(section_texts[3]),
                                   len(section_texts[4]),
                                   len(section_texts[5]),
                                   item]


### Performance check: 1) malstructured dataframe vs 2) ideal dataframe 

이번엔 각 ngram feature 하나하나가 한 row에 저장될 수 있도록 dataframe을 저장해보았다. 

이렇게 하면 이론적으로는 1)의 경우보다 더 이상적인 dataframe structure이다. 하지만 퍼포먼스를 확인했을 때:

- Wall time: 25.1s
- 필수 columns memory sum: 139711528
- ngrams memory: 252397

wall time도 늘어났고, 무엇보다 메모리를 10배 이상 차지한다는 것을 확인할 수 있다. 

어차피 전체 corpus에서 각 ngram의 hawkish/dovish/total 빈도수를 확인하는 것이 목표기 때문에 이를 pandas operation으로 바로 수행하지 못하고 리스트를 처리하거나 str을 처리하는데 오버헤드가 추가적으로 발생한다고 해도 메모리를 절약하는 것이 더 나은 선택일 것이다. 

df.memory_usage(index=True, deep=True)

df[['date', 'minutes', 'Economic Situation', 'Foreign Currency','ngrams']].memory_usage(index=True, deep=True).sum()

df.ngrams

### main function에서 현재 분석에 쓰이지 않는 column은 제거

남긴 columns: `date`, `minutes`, `Foreign Currency`, `Financial Markets`

def preprocessing(source_folder="./data_files/BOK_minutes/pdf2txt/", output_file="./data_files/BOK_minutes/pkl/") :
    # 지정 폴더 내 파일 목록 조회 (파일만)
    txt_files = [f for f in os.listdir(source_folder) if os.path.isfile(os.path.join(source_folder, f))]
    txt_files = [f for f in txt_files if f[-4:] == '.txt']
    txt_files.sort()
    df = pd.DataFrame(columns=['date', 'minutes'])
    df["Foreign Currency"] = ""
    df["Financial Markets"] = ""
    df['ngrams'] =""
    

    for txt_file in txt_files :
        try :
            with open(source_folder + txt_file, 'r',encoding='utf-8') as f :
                txt = f.read()
                sections, section_texts = worker.preprocess_minutes(txt)

                print(txt_file)

                ngrams = []
                for i, section_text in enumerate(section_texts[1:3]):
                    print("doing #{} iteration...".format(i))
                    with mp.Pool(processes=6) as p:
                        mp_result = p.map(worker.text2ngram, section_text)
                    for i in mp_result:
                        ngrams += i
                        
            df.loc[len(df)] = [
                txt_file.split('_-_')[0], # date
                txt, # minutes
                '@@@'.join(section_texts[1]), # Foreign Currency
                '@@@'.join(section_texts[2]), # Financial Markets
                ngrams, # ngrams (list)
            ]
            
        except :
            print('Error occurred at {}'.format(txt_file))
            print('Error: ', sys.exc_info()[0])
            pass
    
    df.to_pickle(output_file + 'minutes_ngram.pkl')

BOK 의사록의 경우, 20분 정도 걸린다. 

%%time

preprocessing()

df = pd.read_pickle('./data_files/BOK_minutes/pkl/minutes_ngram.pkl')
df.head()

# 2. Naver Finance Bond Reports

source_dir="./data_files/NAVER_bond_reports/pdf2txt/"

txt_files = [f for f in os.listdir(source_dir) if os.path.isfile(os.path.join(source_dir, f))]
txt_files = [f for f in txt_files if f[-4:] == '.txt']
txt_files.sort()

len(txt_files)

### Preprocessing bond reports 

- 다음과 같은 문제가 있었다. 
    - pdf 2 txt로 변환되면서한 문장 내에서 줄바꿈이 되어도 제대로 인식하지 못한다. 즉, 문장이 분리되었다. 
    - line별로 ngram을 돌리면 지나치게 시간이 많이 걸렸다. (보고서 1개에 Wall time 약 4분)
- 다음과 같이 전처리 하였다. 
    - `readlines()` 시 의미없이 시작부분에 들어간 `\n` (new line character)를 제거했다. 
    - 남은 부분을 모두 `strip`하고 `' '.join(txt)`로 이어붙여 하나의 문서로 만들었다. 
- 이로 인해 기대되는 효과는
    - 대부분의 문장이 pdf 2 txt 로 변환되며 떨어졌는데, 이를 한 문장으로 이어줄 수 있다. 
        - 하지만 문장이 단어단위로 줄넘김되지 않았을 경우는 `' '` 공백이 들어가 단어가 이상하게 연결될 수도 있다. 
    - 하나의 document로 만들어 ngram을 돌림으로써 연산시간을 단축할 수 있다. 

def preprocessing_bond_reports(source_dir="./data_files/NAVER_bond_reports/pdf2txt/", output_dir="./data_files/NAVER_bond_reports/pkl/"):
    txt_files = [f for f in os.listdir(source_dir) if os.path.isfile(os.path.join(source_dir, f))]
    txt_files = [f for f in txt_files if f[-4:] == '.txt']
    txt_files.sort()
    print("There are {} reports".format(len(txt_files)))
    
    all_reports_text = []

    for txt_file in txt_files:
        with open(source_folder + txt_file, 'r',encoding='utf-8') as f :
            txt = f.readlines()
            txt = [re.sub('^[ ]+', '', x) for x in txt]
            txt = [x.strip() for x in txt if x!='\n']
            all_reports_text.append(' '.join(txt))

    with mp.Pool(processes=6) as p:
        mp_result = p.map(worker.text2ngram, all_reports_text)
    print("There are {} ngram results".format(len(mp_result)))
          
    reports_ngram_dic_list = []

    for row in zip(txt_files, mp_result):
        tmp_file_name = row[0].split('_')

        reports_ngram_dic = {}
        reports_ngram_dic['date'] = tmp_file_name[0]
        reports_ngram_dic['title'] = tmp_file_name[1]
        reports_ngram_dic['ngrams'] = row[1]
        reports_ngram_dic_list.append(reports_ngram_dic)
    
    df = pd.DataFrame(reports_ngram_dic_list)
    df.to_pickle(output_dir + 'bond_reports_ngram.pkl')
        
        

%%time

preprocessing_bond_reports()
