==> 04_polarity_dictionary.ipynb <==
# Creating Polarity Dictionary

극성사전을 만든다. 

우선 한국은행 의사록, 네이버 금융 증권사 채권보고서, 뉴스 데이터에서 각 일자, 문서별로 ngram 리스트를 뽑아낸 데이터를 불러온다. 

추가적으로 일일 콜금리 데이터를 불러온다. 이 때, 콜금리 데이터는 30일 전과 비교하여 당일 콜금리가 증가했는지 감소했는지 레이블 되어있다. 이를 토대로 금리가 올랐다면 hawkish, 내렸다면 dovish로 가정한다. 

의사록, 채권보고서, 뉴스 데이터 모두 date, ngrams 컬럼만 뽑아낸 뒤 일자별로 콜금리 증감 여부를 merge한다. 

이제 어떤 ngrams (list)가 hawkish하고 dovish한지 레이블 되었기 때문에 일자 상관없이 각 ngram 하나하나의 hawkish 빈도와 dovish 빈도, 그리고 total (= hawkish + dovish) 빈도를 표시한다. 

이제 Naive Bayes Classifier의 가정 하에 각 ngram의 베이즈확률을 구하고 Polarity Score을 구할 수 있다. 

import pandas as pd
import numpy as np

import multiprocessing as mp

## Open data and merge them 

call_rates_df = pd.read_pickle('./data_files/BOK_call_rates.pkl')
call_rates_df.head()

minutes_df = pd.read_pickle('./data_files/BOK_minutes/pkl/minutes_ngram.pkl')
minutes_df.date = pd.to_datetime(minutes_df.date, format='%Y-%m-%d')
minutes_df = minutes_df
minutes_df.head()

len(minutes_df)

bond_report_df = pd.read_pickle('./data_files/NAVER_bond_reports/pkl/bond_reports_ngram.pkl')
bond_report_df.date = pd.to_datetime(bond_report_df.date, format='%Y-%m-%d')
bond_report_df.head()

len(bond_report_df)

ngram_freq_df = minutes_df[['date', 'ngrams']]
ngram_freq_df = ngram_freq_df.append(bond_report_df[['date', 'ngrams']], ignore_index=True)
ngram_freq_df.head()

len(ngram_freq_df)

ngram_freq_df = ngram_freq_df.merge(call_rates_df[['date', 'change']], how='left', on='date', )
print(len(ngram_freq_df))
ngram_freq_df.head()

## Get unique ngrams

all_ngrams = []

for li in ngram_freq_df.ngrams:
    all_ngrams += li

unique_ngrams = set(all_ngrams)

## Count unique ngrams' frequency by label (hawkish/dovish)

ngram_freq_dovish_df = ngram_freq_df[ngram_freq_df.change == -1]
ngram_freq_hawkish_df = ngram_freq_df[ngram_freq_df.change == 1]

unique_ngrams_freq_dic = {}

for element in unique_ngrams:
    unique_ngrams_freq_dic[element] = {'hawkish':0, 'dovish':0}

for ngram_list in ngram_freq_dovish_df.ngrams:
    for ngram in ngram_list:
        unique_ngrams_freq_dic[ngram]['dovish'] += 1
    

for ngram_list in ngram_freq_hawkish_df.ngrams:
    for ngram in ngram_list:
        unique_ngrams_freq_dic[ngram]['hawkish'] += 1
    

unique_ngrams_freq_df = pd.DataFrame.from_dict(unique_ngrams_freq_dic, orient='index', columns=['hawkish', 'dovish'])
unique_ngrams_freq_df['total'] = unique_ngrams_freq_df.hawkish + unique_ngrams_freq_df.dovish
unique_ngrams_freq_df.head()

### Drop ngram total freq < 15

unique_ngrams_freq_df = unique_ngrams_freq_df[unique_ngrams_freq_df.total > 15]

## Calculate Bayes probabilities & create polarity dictionary

sum_hawkish = unique_ngrams_freq_df.hawkish.sum()
sum_dovish = unique_ngrams_freq_df.dovish.sum()
sum_total = sum_hawkish + sum_dovish

unique_ngrams_freq_df['prob_feature_given_hawkish'] = unique_ngrams_freq_df.hawkish/sum_hawkish
unique_ngrams_freq_df['prob_feature_given_dovish'] = unique_ngrams_freq_df.dovish/sum_dovish
unique_ngrams_freq_df['polarity_score'] = unique_ngrams_freq_df.prob_feature_given_hawkish/unique_ngrams_freq_df.prob_feature_given_dovish

unique_ngrams_freq_df.polarity_score = unique_ngrams_freq_df.polarity_score.replace([np.inf, -np.inf], np.nan)
unique_ngrams_freq_df.dropna(inplace=True)

unique_ngrams_freq_df.head()

### Drop ngrams in the "grey area". i.e, neutral ngrams. 

intensity threshold = 1.3

intensity_threshold = 1.3

unique_ngrams_freq_df = unique_ngrams_freq_df[(unique_ngrams_freq_df.polarity_score >= intensity_threshold) | (unique_ngrams_freq_df.polarity_score <= 1/intensity_threshold)]

unique_ngrams_freq_df.to_pickle('./data_files/results/polarity_dictionary.pkl')
